refactor(editor): Fold the design-system theme build into the one vite config (no-changelog)
Review feedback: a second config file was the wrong shape. The theme pass now
runs from `builder.buildApp` in `vite.config.mts`, so `pnpm build` is plain
`vite build` again and `package.json` is untouched.

It still needs its own pass rather than a second `lib.entry`: the entry is
`src/css/index.scss`, not a JS module, and a second entry forces code splitting,
which the UMD output cannot do. Vite's `environments` API was tried first and
fails on a CSS-only entry (`vite:css-post` reads `viteMetadata` off an undefined
chunk), so the pass is an `InlineConfig` in the same file.

`base: './'` stays scoped to that pass. Measured: the library pass emits no
assets at all and its CSS carries no non-`data:` `url()`, so the setting would
have been a no-op there either way.

Also renames the emitted component stylesheet `design-system.css` to the
conventional `style.css`, which pairs with `theme.css` and matches what the
later `exports` map wants. Free to do: the old name appears nowhere in the repo,
no `style` field declares it, and every internal consumer aliases the package to
`src`. Content is unchanged — `style.css` has the same sha256 as `master`'s
`design-system.css`, as do both JS bundles.
